Fix compile errors I hit when building org.libsdl in source2 (part 2 of 2) @saml

+++ android-project/app/src/main/java/org/libsdl/app/SDLActivity.java
@@ -651,7 +651,7 @@ public class SDLActivity extends Activity implements View.OnSystemUiVisibilityCh
             }
             case COMMAND_CHANGE_SURFACEVIEW_FORMAT:
             {
-                int format = ((int)msg.obj);
+                int format = ((Integer) msg.obj).intValue();
                 int pf;
 
                 if (SDLActivity.mSurface == null) {
